Table of active substance names extracted from all years starting 2011

Please note that products whose authorisation has expired remain in the files until the end of the period during which use by the end user is still permitted. Therefore, the latest year given in this dataset means the latest year where a use up period ('expirationDeadline') of at least one product containing the substance had not expired in the beginning of the respective year.

Usage

srppp_active_substances_merged

Format

A tibble::tibble

Details

Please see srppp_substances and srppp_substances_merged for details on the different primary keys used in these tables.

Note

Note that a substance occurring in the register in a certain year does not necessarily mean that a product containing the substance was authorized. In particular, a number of products with such comments are present in the register used as representative for 2011.

See also

Examples

print(srppp_active_substances_merged, n = Inf)
#> # A tibble: 648 × 9
#>     pk_v1 pk_v2     earliest latest iupac substance_de substance_fr substance_it
#>     <int> <chr>        <int>  <int> <chr> <chr>        <chr>        <chr>       
#>   1     0 NA            2011   2012 (±)-… Propiconazo… "propiconaz… Propiconazo…
#>   2     3 6F14D297…     2011   2026 1-na… 1-Naphthyla… "1-naphtyl … Acido 1-naf…
#>   3     4 CED72E0E…     2011   2026 (2,4… 2,4-D        "2,4-D"      2,4-D       
#>   4     8 B12482B3…     2011   2026 Aver… Abamectin    "abamectine" Abamectina  
#>   5    10 NA            2011   2011 O,S-… Acephate     "acéphate"   Acefato     
#>   6    13 FEC03074…     2011   2026 2-ch… Aclonifen    "aclonifène" Aclonifen   
#>   7    18 NA            2011   2011 ethy… Alanycarb    "alanycarbe" Alanycarb   
#>   8    19 NA            2011   2017 2-me… Aldicarb     "aldicarbe"  Aldicarb    
#>   9    24 NA            2011   2022 (S)-… zeta-Cyperm… "zeta-cyper… zeta-Ciperm…
#>  10    26 AB9B02C4…     2011   2026 3-(4… Amidosulfur… "amidosulfu… Amidosulfur…
#>  11    32 NA            2011   2011 4,6-… Anilazin     "anilazine"  Anilazina   
#>  12    35 AEE4CE26…     2011   2026 Meth… Asulam       "asulame"    Asulame     
#>  13    37 NA            2011   2011 1-[[… Azaconazole  "azaconazol… Azaconazolo 
#>  14    44 NA            2011   2011 1-(4… Prosulfuron  "prosulfuro… Prosulfuron 
#>  15    48 NA            2011   2011 meth… Benomyl      "benomyl"    Benomil     
#>  16    50 82D8F548…     2011   2026 3-is… Bentazon     "bentazone"  Bentazone   
#>  17    51 NA            2011   2011 3-ch… Benzoximat   "benzoximat… Benzossimato
#>  18    52 NA            2011   2018 meth… Bifenox      "bifénox"    Bifenox     
#>  19    70 NA            2011   2022 3,5-… Bromoxynil   "bromoxynil" Bromoxinil  
#>  20    71 BD9B7DE0…     2011   2026 5-bu… Bupirimate   "bupirimate" Bupirimate  
#>  21    72 64CAB174…     2011   2026 (Z)-… Buprofezin   "buprofézin… Buprofezin  
#>  22    75 NA            2011   2011 N-se… Butralin     "butraline"  Butraline   
#>  23    79 05EE341B…     2011   2026 N-(t… Captan       "captan"     Captano     
#>  24    81 NA            2011   2018 meth… Carbendazim  "carbendazi… Carbendazim 
#>  25    82 NA            2011   2018 (R)-… Carbetamide  "carbétamid… Carbetamide 
#>  26    83 NA            2011   2013 2,3-… Carbofuran   "carbofuran… Carbofuran  
#>  27    85 NA            2011   2013 2,3-… Carbosulfan  "carbosulfa… Carbosulfan 
#>  28    86 NA            2011   2014 5,6-… Carboxin     "carboxine"  Carbossina  
#>  29    90 NA            2011   2013 2-ch… Chlorfenvin… "chlorfenvi… Clorfenvinf…
#>  30    92 NA            2011   2021 5-am… Chloridazon  "chloridazo… Cloridazon  
#>  31    94 NA            2011   2020 2-ch… Chlormequat… "chlormequa… Clormequat …
#>  32    99 NA            2011   2020 tetr… Chlorothalo… "chlorothal… Clorotaloni…
#>  33   101 NA            2011   2021 O,O-… Chlorpyrifos "chlorpyrif… Clorpirifos 
#>  34   102 NA            2011   2020 O,O-… Chlorpyrifo… "chlorpyrip… Clorpirifos…
#>  35   106 62EEA366…     2011   2026 (5RS… Clethodim    "cléthodime" Clethodim   
#>  36   107 NA            2011   2025 3,6-… Clofentezine "clofentézi… Clofentezine
#>  37   108 FCEE064F…     2011   2026 2-(2… Clomazone    "clomazone"  Clomazone   
#>  38   110 FC3EB8FF…     2011   2026 3,6-… Clopyralid   "clopyralid… Clopiralid  
#>  39   112 NA            2011   2015 NA    Cyanamid     "cyanamide"  Cianammide  
#>  40   113 NA            2011   2012 tric… Cyhexatin    "cyhexatine" Ciexatin    
#>  41   114 9D9A5C3D…     2011   2026 1-(2… Cymoxanil    "cymoxanil"  Cimoxanil   
#>  42   115 9BA42DE8…     2011   2026 (RS)… Cypermethrin "cyperméthr… Cipermetrina
#>  43   116 NA            2011   2023 (2RS… Cyproconazo… "cyproconaz… Ciproconazo…
#>  44   117 NA            2011   2013 N-cy… Cyromazin    "cyromazine" Ciromazina  
#>  45   119 81778438…     2011   2026 N-di… Daminozide   "daminozide" Daminozide  
#>  46   121 NA            2011   2022 ethy… Desmedipham  "desmédipha… Desmedifam  
#>  47   123 NA            2011   2013 O,O-… Diazinon     "diazinon"   Diazinone   
#>  48   124 738BFE0C…     2011   2026 3,6-… Dicamba      "dicamba"    Dicamba     
#>  49   125 NA            2011   2015 2,6-… Dichlobenil  "dichlobeni… Diclobenil  
#>  50   127 NA            2011   2011 (R)-… Dichlorprop… "dichlorpro… Diclorprop-P
#>  51   128 NA            2011   2012 2,2-… Dichlorvos … "dichlorvos… Diclorvos (…
#>  52   135 NA            2011   2023 1-(4… Diflubenzur… "diflubenzu… Diflubenzur…
#>  53   138 NA            2011   2022 O,O-… Dimethoate   "diméthoate" Dimetoato   
#>  54   152 514F42F1…     2011   2026 1-do… Dodine       "dodine"     Dodina      
#>  55   155 NA            2011   2013 (1,4… Endosulfan   "endosulfan" Endosulfan  
#>  56   162 1F6392CF…     2011   2026 2-Ch… Ethephon     "éthéphon"   Etefon      
#>  57   165 5529C655…     2011   2026 (±)-… Ethofumesate "éthofumésa… Etofumesate 
#>  58   171 NA            2011   2017 4-te… Fenazaquin   "fenazaquin" Fenazaquin  
#>  59   172 NA            2011   2011 O,O-… Fenitrothion "fénitrothi… Fenitrotion 
#>  60   174 NA            2011   2022 ethy… Fenoxycarb   "fénoxycarb… Fenoxicarb  
#>  61   176 NA            2011   2021 (+/-… Fenpropimor… "fenpropimo… Fenpropimorf
#>  62   177 NA            2011   2011 trip… Fentin acet… "acétate de… Fentin acet…
#>  63   188 NA            2011   2018 bis(… Flusilazole  "flusilazol… Flusilazol  
#>  64   190 683783D6…     2011   2026 N-(t… Folpet       "folpet"     Folpet      
#>  65   196 37DDAF9E…     2011   2026 etho… Fosetyl      "fosétyl"    Fosetil     
#>  66   199 7B9F385E…     2011   2026 N-(p… Glyphosat    "glyphosate" Glifosato   
#>  67   200 NA            2011   2011 Mixt… Guazatine    "guazatine"  Guazatina   
#>  68   203 NA            2011   2011 (RS)… Hexaconazole "hexaconazo… Esaconazolo 
#>  69   205 A4D10A95…     2011   2026 (4RS… Hexythiazox  "hexythiazo… Essitiazox  
#>  70   230 NA            2011   2020 3-(3… Iprodione    "iprodione"  Iprodione   
#>  71   269 EAD03721…     2011   2026 2-(4… Nicosulfuron "nicosulfur… Nicosulfuron
#>  72   276 NA            2011   2022 3,5-… Oryzalin     "oryzalin"   Orizalin    
#>  73   283 NA            2011   2017 2-ch… Oxyfluorfen  "oxyfluorfe… Oxifluorfen 
#>  74   284 42BD4717…     2011   2026 (2RS… Paclobutraz… "paclobutra… Paclobutraz…
#>  75   286 207EF3C9…     2011   2026 1-(2… Penconazole  "penconazol… Penconazolo 
#>  76   287 F0425D85…     2011   2026 N-(1… Pendimethal… "pendimétha… Pendimetalin
#>  77   289 NA            2011   2011 3-ph… Permethrin   "permethrin… Permetrina  
#>  78   290 NA            2011   2013 NA    Mineralöl /… "huile miné… Oli mineral…
#>  79   291 03CDB89A…     2011   2026 meth… Phenmedipham "phenmediph… Fenmedifam  
#>  80   295 NA            2011   2013 S-6-… Phosalone    "phosalone"  Fosalone    
#>  81   299 CDC2B69D…     2011   2026 4-am… Picloram     "piclorame"  Picloram    
#>  82   301 A2DD5346…     2011   2026 2-di… Pirimicarb   "pirimicarb… Pirimicarb  
#>  83   303 NA            2011   2017 O,O-… Pirimiphos-… "pirimiphos… Pirimifos-m…
#>  84   305 NA            2011   2023 N-pr… Prochloraz   "prochloraz" Procloraz   
#>  85   307 NA            2011   2011 O-4-… Profenofos   "profenofos" Profenofos  
#>  86   310 NA            2011   2013 2-ch… Propachlor   "propachlor… Propaclor   
#>  87   311 1D9BCAC4…     2011   2026 Prop… Propamocarb  "propamocar… Propamocarb 
#>  88   312 3FCB2A84…     2011   2026 2-is… Propaquizaf… "propaquiza… Propaquizaf…
#>  89   318 NA            2011   2017 poly… Propineb     "propinèbe"  Propineb    
#>  90   323 7639690D…     2011   2026 Pyre… Pyrethrine   "pyréthrine… Piretrine   
#>  91   338 D95F01F3…     2011   2026 sulf… Schwefel     "soufre"     Zolfo       
#>  92   344 714F6C9B…     2013   2026 2,3,… Tefluthrin   "téfluthrin… Teflutrin   
#>  93   360 NA            2011   2011 2-[1… Tralkoxydim  "tralkoxydi… Tralcossidim
#>  94   451 0E40457A…     2011   2026 3,5-… Dazomet (DM… "dazomet (D… Dazomet (DM…
#>  95   453 37D99E20…     2011   2026 (S)-… Deltamethrin "deltaméthr… Deltametrina
#>  96   529 NA            2011   2019 4-(4… Fenbuconazo… "fenbuconaz… Fenbuconazo…
#>  97   564 A08E209B…     2011   2026 1-(4… Flazasulfur… "flazasulfu… Flazasulfur…
#>  98   642 NA            2011   2013 7-ch… Heptenophos  "heptenopho… Eptenofos   
#>  99   701 NA            2011   2025 zinc… Metiram      "métirame"   Metiram     
#> 100   706 D8A1C403…     2011   2026 4-am… Metribuzin   "métribuzin… Metribuzin  
#> 101   788 NA            2014   2014 3-(4… Thifensulfu… "thifensulf… Tifensulfur…
#> 102   793 NA            2011   2021 O-2,… Tolclofos-m… "tolclofos-… Tolclofos-m…
#> 103   795 NA            2011   2020 (1RS… Triadimenol  "triadiméno… Triadimenol 
#> 104   798 NA            2011   2023 7-ch… Triazoxid    "triazoxide" triazoxide  
#> 105   799 NA            2011   2025 2-[4… Tribenuron   "tribénuron" Tribenuron  
#> 106   801 NA            2011   2011 dime… Trichlorfon  "trichlorfo… Triclorfon  
#> 107   803 NA            2011   2017 (E)-… Triflumizole "triflumizo… Triflumizolo
#> 108   806 NA            2011   2011 O,O-… Vamidothion  "vamidothio… Vamidotion  
#> 109   808 NA            2011   2017 Ento… Verticilliu… "Verticilli… Verticilliu…
#> 110   809 NA            2011   2012 (RS)… Vinclozolin  "vinclozoli… Vinclozolin 
#> 111   813 NA            2011   2025 2-(1… 2-(1-naphth… "2-(1-napht… 2-(1-naftil…
#> 112   815 NA            2011   2011 O,O-… Parathion-m… "parathion-… Paration-me…
#> 113   832 0C22658C…     2011   2026 Para… Encarsia fo… "Encarsia f… Encarsia fo…
#> 114   833 NA            2011   2017 Raub… Amblyseius … "Amblyseus … Amblyseius …
#> 115   834 NA            2011   2025 Ento… Cydia pomon… "Cydia pomo… Cydia pomon…
#> 116   835 EDCC8B9C…     2011   2026 Raub… Amblyseius … "Amblyseius… Amblyseius …
#> 117   836 53F9C24F…     2011   2026 Para… Aphelinus a… "Aphelinus … Aphelinus a…
#> 118   837 BA4F77C9…     2015   2026 Para… Aphidius ma… "Aphidius m… Aphidius ma…
#> 119   838 9E22008E…     2011   2026 Gall… Aphidoletes… "Aphidolete… Aphidoletes…
#> 120   839 858B09AE…     2011   2026 Räub… Cryptolaemu… "Cryptolaem… Cryptolaemu…
#> 121   840 62CEFAC8…     2011   2026 Para… Dacnusa sib… "Dacnusa si… Dacnusa sib…
#> 122   841 25046794…     2011   2026 Para… Diglyphus i… "Diglyphus … Diglyphus i…
#> 123   842 NA            2011   2017 Ento… Heterorhabd… "Heterorhab… Heterorhabd…
#> 124   843 045FC758…     2011   2026 Stre… Streptomyce… "Streptomyc… Streptomyce…
#> 125   844 NA            2011   2017 Oriu… Orius insid… "Orius insi… Orius insid…
#> 126   845 1C3330C0…     2011   2026 Räub… Orius laevi… "Orius laev… Orius laevi…
#> 127   846 F10ED28E…     2011   2026 Räub… Orius majus… "Orius maju… Orius majus…
#> 128   847 9C25296A…     2011   2026 Raub… Phytoseiulu… "Phytoseiul… Phytoseiulu…
#> 129   849 56999C96…     2011   2026 Ento… Beauveria b… "Beauveria … Beauveria b…
#> 130   851 FBE355DE…     2011   2026 Ento… Schalenwick… "virus de l… Virus della…
#> 131   852 18364CCD…     2011   2026 Ento… Steinernema… "Steinernem… Steinernema…
#> 132   853 5363C4A2…     2011   2026 Ento… Steinernema… "Steinernem… Steinernema…
#> 133   854 0FD8F60C…     2011   2026 Para… Trichogramm… "Trichogram… Trichogramm…
#> 134   858 NA            2011   2017 2-Et… Chalcogran   "Chalcogran" Chalcogran  
#> 135   863 NA            2011   2025 3,5-… Propyzamid   "propyzamid… Propizamide 
#> 136   867 NA            2011   2011 3-(4… Monolinuron  "monolinuro… Monolinuron 
#> 137   868 NA            2011   2021 4-[h… Glufosinate  "glufosinat… Glufosinate 
#> 138   869 NA            2011   2020 3-(4… Isoproturon  "isoproturo… Isoproturon 
#> 139   870 NA            2011   2011 1-[2… Triasulfuron "triasulfur… Triasulfuron
#> 140   872 32AD9072…     2011   2026 ethy… Fenoxaprop-… "fénoxaprop… Fenossaprop…
#> 141   873 D6906EB7…     2011   2026 2-ch… Dimethachlor "diméthachl… Dimetaclor  
#> 142   875 7AD9E7AE…     2011   2026 4-am… Fluroxypyr   "fluroxypyr" Fluroxypyr  
#> 143   876 9B96406A…     2011   2026 3-cy… Lenacil      "lénacile"   Lenacil     
#> 144   877 E86FD4D8…     2011   2026 (R)-… Mecoprop-P   "mécoprop-P" Mecoprop-P  
#> 145   878 NA            2011   2018 4-hy… Ioxynil      "ioxynil"    Ioxinil     
#> 146   879 E3F7BA6E…     2011   2026 (4-c… MCPA         "MCPA"       MCPA        
#> 147   880 7BF196DB…     2011   2026 4-am… Metamitron   "métamitron" Metamitron  
#> 148   881 92A01916…     2011   2026 3-(3… Chlorotolur… "chlortolur… Clorotoluron
#> 149   882 NA            2011   2013 (RS)… Dimethenamid "diméthénam… Dimetenamid 
#> 150   883 NA            2011   2022 2-p-… Myclobutanil "myclobutan… Miclobutanil
#> 151   884 NA            2011   2012 FeO4… Eisen-II-Su… "sulfate de… Solfato fer…
#> 152   885 NA            2011   2012 NA    Calciumcyan… "cyanamide … Calciociana…
#> 153   886 NA            2011   2012 NA    Alkyl-dimet… "chlorure d… Alchil-dime…
#> 154   887 NA            2011   2011 2',4… Pyrifenox    "pyrifénox"  Pirifenox   
#> 155   890 F26735AB…     2011   2026 1-(4… Rimsulfuron  "rimsulfuro… Rimsulfuron 
#> 156   891 NA            2011   2016 2-(2… Sulcotrione  "sulcotrion… Sulcotrione 
#> 157   893 D4C6CB2D…     2011   2026 (RS)… Tebuconazole "tébuconazo… Tebuconazolo
#> 158   894 112F60C6…     2011   2026 cis,… Difenoconaz… "difénocona… Difenoconaz…
#> 159   895 NA            2011   2020 Isop… Chlorpropha… "chlorproph… Clorprofam …
#> 160   896 12F5B2CC…     2011   2026 3-ch… Fluazinam    "fluazinam"  Fluazinam   
#> 161   897 NA            2011   2025 copp… Kupfer (als… "cuivre (so… Rame (sotto…
#> 162   898 NA            2011   2025 dico… Kupfer (als… "cuivre (so… Rame (sotto…
#> 163   909 NA            2011   2022 mang… Mancozeb     "mancozèbe"  Mancozeb    
#> 164   912 NA            2011   2025 (E,Z… Dimethomorph "diméthomor… Dimetomorf  
#> 165   914 NA            2011   2017 mang… Maneb        "manèbe"     Maneb       
#> 166   917 NA            2011   2022 dime… Thiophanate… "thiophanat… Tiofanato-m…
#> 167   918 3181BC7A…     2011   2026 Fenc… Oleum foeni… "oleum foen… Oleum foeni…
#> 168   920 11AA9D40…     2011   2026 2-(t… Thiabendazo… "thiabendaz… Tiabendazolo
#> 169   921 NA            2011   2021 tetr… Thiram (TMT… "thirame (T… Tiram (TMTD)
#> 170   924 D973ED99…     2011   2026 (±)-… Imazalil     "imazalil"   Imazalil    
#> 171   925 B69935D6…     2011   2026 5-me… Hymexazol    "hymexazole" Hymexazol   
#> 172   926 NA            2011   2011 copp… Kupfer (als… "cuivre (so… Rame (sotto…
#> 173   928 CE067EA3…     2011   2026 (±)-… Triticonazo… "triticonaz… Triticonazo…
#> 174   934 NA            2011   2020 di-1… Pinolene     "pinolène"   Pinolene    
#> 175   935 1B400D81…     2011   2026 Baci… Bacillus th… "Bacillus t… Bacillus th…
#> 176   937 158EE23A…     2014   2026 Baci… Bacillus th… "Bacillus t… Bacillus th…
#> 177   938 NA            2011   2020 NA    Kupfer (als… "cuivre (so… Rame (sotto…
#> 178   939 75729E67…     2011   2026 Baci… Bacillus th… "Bacillus t… Bacillus th…
#> 179   941 NA            2011   2024 Meth… Triflusulfu… "triflusulf… Triflusulfu…
#> 180   942 2ECA23B2…     2011   2026 ethy… Quizalofop-… "quizalofop… Quizalofop-…
#> 181   943 0F29D476…     2011   2026 buty… Fluazifop-P… "fluazifop-… Fluazifop-P…
#> 182   945 NA            2011   2022 (RS)… Lufenuron    "lufénuron"  Lufenuron   
#> 183   946 B89C332B…     2011   2026 (RS)… Fenpropidin  "fenpropidi… Fenpropidin 
#> 184   948 6CE21C18…     2011   2026 N-te… Tebufenozide "tébufénozi… Tebufenozide
#> 185   949 NA            2011   2013 1-[3… Hexaflumuron "hexaflumur… Hexaflumuron
#> 186   950 NA            2011   2025 (E,E… E8E10-12 OH… "E8E10-12 O… E8E10-12 OH…
#> 187   951 NA            2011   2025 (Z)-… Z11-14 Ac    "Z11-14 Ac"  Z11-14 Ac   
#> 188   952 88B71880…     2011   2026 4-cy… Cyprodinil   "cyprodinil" Ciprodinil  
#> 189   955 E972B717…     2011   2026 Para… Aphidius co… "Aphidius c… Aphidius co…
#> 190   957 NA            2011   2011 3-[(… Difethialone "diféthialo… Difethialone
#> 191   959 NA            2011   2011 Natr… Natriumdike… "dikégulac-… Dikegulac-s…
#> 192   960 E5044E08…     2011   2026 Ento… Metarhizium… "Metarhiziu… Metarhizium…
#> 193   961 DC45C48E…     2011   2026 6-hy… Maleinsäure… "Hydrazide … Idrazide ma…
#> 194   962 NA            2011   2015 NA    Kieselsäure… "acide sili… Acido silic…
#> 195   970 3DD73DF0…     2011   2026 Baci… Bacillus th… "Bacillus t… Bacillus th…
#> 196   971 NA            2011   2022 1-(6… Imidacloprid "imidaclopr… Imidacloprid
#> 197   972 526B05DE…     2011   2026 Ento… Beauveria b… "Beauveria … Beauveria b…
#> 198   973 NA            2011   2014 (+/-… Fipronil     "fipronil"   Fipronil    
#> 199   974 BCDB5D6B…     2011   2026 NA    Phasmarhabd… "Phasmarhab… Phasmarhabd…
#> 200   975 NA            2011   2025 (Z)-… Z8-12Ac      "Z8-12Ac"    Z8-12Ac     
#> 201   976 NA            2011   2025 (E)-… E8-12Ac      "E8-12Ac"    E8-12Ac     
#> 202   977 NA            2011   2025 (Z)-… Z8-12OH ((Z… "Z8-12OH ((… Z8-12OH ((Z…
#> 203   980 1A2BDF91…     2011   2026 N-(4… Pyrimethanil "pyriméthan… Pyrimethanil
#> 204   981 NA            2011   2025 N-(4… Mepanipyrim  "mépanipyri… Mepanipyrim 
#> 205   982 2C992C62…     2011   2026 4-(2… Fludioxonil  "fludioxoni… Fludioxonil 
#> 206   984 D7339D7B…     2011   2026 Phle… Phlebia gig… "Phlebia gi… Phlebia gig…
#> 207   986 NA            2011   2017 NA    Lecithin     "lécithine"  Lecitina    
#> 208   988 1F1AC43C…     2011   2026 Leci… Sojalecithin "lécithine … Lecitina di…
#> 209   990 NA            2011   2022 (2RS… Epoxiconazo… "époxiconaz… Epoxiconazol
#> 210   995 NA            2011   2011 oxet… Oxasulfuron  "oxasulfuro… Oxasulfuron 
#> 211   997 NA            2011   2017 Para… Encyrtus le… "Encyrtus l… Encyrtus le…
#> 212   998 FCC3428E…     2011   2026 NA    Leptomastid… "Leptomasti… Leptomastid…
#> 213   999 F2933BD8…     2011   2026 NA    Leptomastix… "Leptomasti… Leptomastix…
#> 214  1000 NA            2011   2022 (E)-… Pymetrozine  "pymétrozin… Pimetrozine 
#> 215  1002 31403F9A…     2011   2026 copp… Kupfer       "cuivre"     Rame        
#> 216  1008 NA            2011   2011 zinc… Zineb        "zinèbe"     Zineb       
#> 217  1009 NA            2011   2011 di-i… Nitrothal -… "nitrothal … Nitrotal-is…
#> 218  1010 NA            2011   2017 8-Hy… Oxychinolin  "oxychinoli… Oxichinolina
#> 219  1015 NA            2011   2017 Baci… Bacillus th… "Bacillus t… Bacillus th…
#> 220  1016 NA            2011   2011 3-me… Butoxycarbo… "butoxycarb… Butossicarb…
#> 221  1019 NA            2011   2011 bis[… Fenbutatin … "fenbutatin… Fenbutatin …
#> 222  1020 NA            2011   2012 S-2,… Methidathion "méthidathi… Metidation  
#> 223  1021 NA            2011   2011 2-me… Mevinphos    "mévinphos"  Mevinfos    
#> 224  1024 NA            2011   2013 (2R,… Rotenon      "roténone"   Rotenone    
#> 225  1025 8D005DE7…     2011   2026 2-(2… Piperonyl b… "pipéronyl … Piperonilbu…
#> 226  1027 F976B504…     2011   2026 para… Paraffinöl   "huile de p… Olio di par…
#> 227  1030 NA            2011   2017 NA    Tensid-Gemi… "mélange d'… Miscela di …
#> 228  1032 NA            2011   2019 biph… Orthophenyl… "orthophény… Ortofenilfe…
#> 229  1034 34CDA556…     2011   2026 r-2,… Metaldehyd   "métaldéhyd… Metaldeide  
#> 230  1035 NA            2011   2013 NA    Leim (glue)  "glu"        Colla       
#> 231  1037 NA            2011   2015 KNO3  Kaliumnitra… "nitrate de… Nitrato di …
#> 232  1039 NA            2011   2011 NA    Kreide und … "craie et h… Gesso e cal…
#> 233  1040 NA            2011   2017 NA    Detergentien "détergents" Detergenti  
#> 234  1041 NA            2011   2017 NA    Aethylengly… "monobuthyl… Etilenglico…
#> 235  1046 NA            2011   2011 1-(b… Benzthiazur… "benzthiazu… Benztiazuron
#> 236  1047 NA            2011   2013 3,5-… Bromoxynil … "bromoxynil… Bromoxinil …
#> 237  1048 NA            2011   2013 4-hy… Ioxynil als… "ioxynil so… Ioxinil sot…
#> 238  1051 NA            2011   2013 tetr… Chlorthal    "chlorthal"  Clortal     
#> 239  1060 NA            2011   2018 3-(3… Linuron      "linuron"    Linuron     
#> 240  1065 NA            2011   2013 NA    Flurenolcar… "acide flur… Acido flure…
#> 241  1067 B58976CD…     2011   2026 4-(4… MCPB         "MCPB"       MCPB        
#> 242  1069 9901EF9E…     2011   2026 2-ch… Metazachlor  "métazachlo… Metazaclor  
#> 243  1070 NA            2011   2011 1-(1… Methabenzth… "méthabenzt… Metabenztia…
#> 244  1071 69CDC31C…     2011   2026 3-(4… Metobromuron "métobromur… Metobromuron
#> 245  1072 NA            2011   2012 3-(3… Metoxuron    "métoxuron"  Metoxuron   
#> 246  1075 A1465223…     2011   2026 O-(6… Pyridate     "pyridate"   Piridato    
#> 247  1076 NA            2011   2022 1,1'… Diquat       "diquat"     Diquat      
#> 248  1080 NA            2011   2015 alph… Trifluralin  "triflurali… Trifluralina
#> 249  1092 NA            2011   2011 (RS)… Cyfluthrin   "cyfluthrin… Ciflutrin   
#> 250  1094 3B31FDB8…     2011   2026 1,5-… Glutaraldeh… "glutaraldé… Glutaraldei…
#> 251  1098 NA            2011   2025 (Z)-… Z9-12 Ac     "Z9-12 Ac"   Z9-12 Ac    
#> 252  1103 NA            2011   2012 2-ch… Alachlor     "alachlore"  Alaclor     
#> 253  1106 7C1C8A05…     2011   2026 2-(1… 1-Naphthyla… "1-naphthyl… 1-naftilace…
#> 254  1107 NA            2011   2016 2-(1… alpha-Napht… "alpha-naph… acido alfa-…
#> 255  1108 NA            2011   2017 Al2O3 Aluminiumox… "oxyde d'al… Ossido di a…
#> 256  1109 002F5C84…     2011   2026 Ehty… Aluminiumfo… "fosétyl-al… Fosetil all…
#> 257  1110 0F6DBED8…     2011   2026 Alum… Aluminiumph… "phosphure … Fosfuro di …
#> 258  1111 NA            2011   2012 N,N'… Amitraz      "amitraze"   Amitraz     
#> 259  1115 NA            2011   2013 -     Baumwachs    "mastic à c… Mastice da …
#> 260  1118 NA            2011   2011 meth… Benalaxyl    "bénalaxyl"  Benalaxil   
#> 261  1121 NA            2011   2011 S,S'… Bensultap    "bensultap"  Bensultap   
#> 262  1122 NA            2011   2022 2-me… Bifenthrin   "bifenthrin… Bifentrin   
#> 263  1123 NA            2011   2011 1-(b… Bitertanol   "bitertanol" Bitertanolo 
#> 264  1126 NA            2011   2013 3-[3… Brodifacoum  "brodifacou… Brodifacoum 
#> 265  1127 NA            2011   2022 3-[3… Bromadiolone "bromadiolo… Bromadiolone
#> 266  1128 NA            2011   2014 isop… Bromopropyl… "bromopropy… Bromopropil…
#> 267  1130 4CD602AC…     2011   2026 Calc… Calciumhydr… "hydroxide … Idrossido d…
#> 268  1131 NA            2011   2022 -     Calciumphos… "phosphure … Fosfuro di …
#> 269  1134 NA            2011   2012 1-me… Chlorbufam   "chlorbufam… Clorbufamio 
#> 270  1137 2C7789D4…     2011   2026 (2R)… Cloquintoce… "cloquintoc… Cloquintoce…
#> 271  1138 9EC4DF79…     2011   2026 (±)-… Cycloxydim   "cycloxydim… Ciclossidim 
#> 272  1139 NA            2011   2012 N-cy… Cycluron     "cycluron"   Cicluron    
#> 273  1142 NA            2011   2014 1-te… Diafenthiur… "diafenthiu… Diafentiuron
#> 274  1145 NA            2011   2018 isop… Diethofenca… "diéthofenc… Dietofencarb
#> 275  1147 90EEFA87…     2011   2026 2’,4… Diflufenican "diflufénic… Diflufenican
#> 276  1148 NA            2011   2013 3-[4… Dimefuron    "dimefuron"  Dimefuron   
#> 277  1149 NA            2011   2011 NA    Dimethyl de… "chlorure d… Cloruro di …
#> 278  1155 63C58A64…     2011   2026 5,10… Dithianon    "dithianon"  Ditianon    
#> 279  1156 NA            2011   2022 3-(3… Diuron       "diuron"     Diuron      
#> 280  1162 66C01E01…     2011   2026 tert… Fenpyroxima… "fenpyroxim… Fenpiroxima…
#> 281  1164 NA            2011   2025 fatt… Fettsäuren … "acides gra… Acidi grass…
#> 282  1168 4A17896B…     2011   2026 (3RS… Flurochlori… "flurochlor… Fluroclorid…
#> 283  1169 NA            2011   2013 Natr… Natriumhexa… "hexafluoro… Esafluorosi…
#> 284  1173 NA            2011   2011 buty… Furathiocarb "furathioca… Furatiocarb 
#> 285  1174 NA            2011   2021 (3S,… Gibberellin… "gibberilli… Gibberellin…
#> 286  1175 NA            2011   2017 (3S,… Gibberellin… "acide gibb… Acido gibbe…
#> 287  1176 NA            2011   2022 (R)-… Haloxyfop-(… "haloxyfop-… Haloxyfop-(…
#> 288  1177 NA            2011   2017 -     Hornmehl     "farine de … Limatura co…
#> 289  1183 NA            2011   2017 -     Kupfer (als… "cuivre (so… Rame (sotto…
#> 290  1184 4631106A…     2011   2026 Bord… Kupfer (als… "cuivre (so… Rame (sotto…
#> 291  1186 NA            2011   2023 copp… Kupfer (als… "cuivre (so… Rame (sotto…
#> 292  1187 1FCE99C4…     2011   2026 a-cy… Lambda-Cyha… "lambda-cyh… Lambda-Cial…
#> 293  1189 F73795A7…     2011   2026 Trim… Magnesiumph… "phosphure … Magnesio fo…
#> 294  1191 NA            2011   2014 3'-i… Mepronil     "mépronil"   Mepronil    
#> 295  1193 NA            2011   2020 4-me… Methiocarb … "méthiocarb… Metiocarb (…
#> 296  1194 NA            2011   2022 S-me… Methomyl     "méthomyl"   Metomil     
#> 297  1195 NA            2011   2011 isop… Methoprene   "méthoprène" Metoprene   
#> 298  1196 C70B01C0…     2011   2026 meth… Metsulfuron… "metsulfuro… Metsulfuron…
#> 299  1197 NA            2011   2023 a ra… alpha-Cyper… "alpha-cype… alfa-Ciperm…
#> 300  1198 NA            2011   2017 -     Mineralstof… "substances… Sali minera…
#> 301  1200 5A6C95DD…     2011   2026 (RS)… Napropamide  "napropamid… Napropamide 
#> 302  1201 NA            2011   2013 -     Natriumfluo… "fluorosili… Fluorosilic…
#> 303  1202 NA            2011   2015 -     Naturharze   "résines na… Resine natu…
#> 304  1204 NA            2011   2025 -     Netz- und H… "mouillants… umidificant…
#> 305  1212 NA            2011   2014 S-2-… Orbencarb    "orbencarb"  Orbencarb   
#> 306  1213 NA            2011   2011 2-me… Oxadixyl     "oxadixyl"   Oxadixil    
#> 307  1216 NA            2011   2017 -     Parfümöl     "huille aro… Oli aromati…
#> 308  1217 NA            2011   2022 1-(4… Pencycuron   "pencycuron" Pencicuron  
#> 309  1218 0AB13A00…     2011   2026 prop… Clodinafop-… "clodinafop… Clodinafop-…
#> 310  1220 8654D583…     2011   2026 Prop… Propamocarb… "propamocar… Propamocarb…
#> 311  1221 517E3519…     2011   2026 S-be… Prosulfocarb "prosulfoca… Prosulfocarb
#> 312  1224 4CECCC3C…     2011   2026 Quas… Quassiaextr… "extrait de… Estratto di…
#> 313  1226 333A57AE…     2011   2026 rape… Rapsöl       "huile de c… Olio di col…
#> 314  1227 NA            2011   2011 -     Rebenverede… "cire à gre… Cera per in…
#> 315  1228 NA            2011   2016 4-am… Fluroxypyr … "fluroxypyr… Fluroxypyr …
#> 316  1233 NA            2011   2024 Sili… Siliciumoxyd "oxyde de s… Ossido di s…
#> 317  1234 NA            2011   2011 6-ch… Simazine     "simazine"   Simazina    
#> 318  1236 NA            2011   2011 N-(P… Sulfosat     "sulfosate"  Sulfosate   
#> 319  1239 F130852A…     2011   2026 N-(4… Tebufenpyrad "tébufenpyr… Tebufenpirad
#> 320  1241 NA            2011   2017 1-(3… Teflubenzur… "téflubenzu… Teflubenzur…
#> 321  1242 NA            2011   2013 3-te… Terbacil     "terbacile"  Terbacil    
#> 322  1243 NA            2011   2013 S-te… Terbufos     "terbufos"   Terbufos    
#> 323  1245 736564C2…     2011   2026 N2-t… Terbuthylaz… "terbuthyla… Terbutilazi…
#> 324  1248 9E2CF77C…     2011   2026 meth… Thifensulfu… "thifensulf… Tifensulfur…
#> 325  1249 NA            2011   2012 N,N-… Thiocyclam … "thiocyclam… Tiociclamid…
#> 326  1251 NA            2011   2017 -     Tierkörperm… "farine ani… Farine anim…
#> 327  1252 NA            2011   2012 TiO … Titanoxyd    "oxyde de t… Ossido di t…
#> 328  1253 B5879942…     2011   2026 meth… Tribenuron-… "tribénuron… Tribenuron-…
#> 329  1254 NA            2011   2025 3,5,… Triclopyr    "triclopyr"  Triclopir   
#> 330  1256 NA            2011   2012 N,N'… Triforine    "triforine"  Triforine   
#> 331  1257 F893B128…     2011   2026 Ethy… Trinexapac-… "trinexapac… Trinexapac-…
#> 332  1258 NA            2011   2017 -     Vinylcopoly… "copolymère… Copolimeri …
#> 333  1259 NA            2011   2017 Viny… Vinylesterp… "polymère d… Polimeri di…
#> 334  1260 F8EF3D5B…     2011   2026 zinc… Ziram        "zirame"     Ziram       
#> 335  1261 9F9A7FB6…     2011   2026 imen… Metaphycus … "Metaphycus… Metaphycus …
#> 336  1263 76CB859A…     2011   2026 Ethy… Carfentrazo… "carfentraz… Carfentrazo…
#> 337  1264 B4D856A9…     2011   2026 Raub… Stratiolael… "Stratiolae… Stratiolael…
#> 338  1265 DB475BCA…     2011   2026 dime… Azadirachti… "azadiracht… Azadiractin…
#> 339  1267 5AB5D02D…     2011   2026 1-(4… Sulfosulfur… "sulfosulfu… Sulfosulfur…
#> 340  1268 05177A88…     2019   2026 7-ch… Quinmerac    "quinmérac"  Quinmerac   
#> 341  1270 NA            2011   2012 2-(4… Cyanazine    "cyanazines" Cianazina   
#> 342  1271 BCC0693F…     2011   2026 Ampe… Ampelomyces… "Ampelomyce… Ampelomyces…
#> 343  1274 445D3713…     2011   2026 poly… Alkoholetho… "éthoxylate… Etossilato …
#> 344  1276 NA            2011   2021 2-am… Quinoclamine "quinoclami… Chinoclamina
#> 345  1277 DF3F9C6A…     2013   2026 pota… Kaliumphosp… "phosphonat… potassio fo…
#> 346  1282 NA            2011   2018 d-2-… d-Carvon     "d-carvon"   d-Carvone   
#> 347  1283 46D4F9FD…     2011   2026 (S)-… Ipsdienol    "ipsdiénol"  Ipsdienol   
#> 348  1284 NA            2011   2017 Deca… Decadiencar… "ester méth… Estere meti…
#> 349  1285 NA            2011   2020 NA    Ätherische … "huiles ess… Oli essenzi…
#> 350  1289 7CC210AC…     2011   2026 [S-(… (S)-cis-Ver… "(S)-cis-Ve… (S)-cis-Ver…
#> 351  1291 NA            2011   2017 Diam… Aethoxylier… "octylphéno… Etossilato …
#> 352  1298 NA            2011   2025 7-S-… Acibenzolar… "acibenzola… Acibenzolar…
#> 353  1299 6AFF308C…     2011   2026 meth… Metalaxyl-M  "métalaxyle… Metalaxil-M 
#> 354  1301 21475125…     2011   2026 (1RS… Metconazole  "metconazol… Metconazolo 
#> 355  1303 NA            2011   2017 octa… Octansäure … "acide octa… Acido ottan…
#> 356  1304 NA            2011   2025 copp… Kupfer (als… "cuivre (so… Rame (sotto…
#> 357  1305 NA            2011   2025 (7E,… E7Z9-12 Ac   "E7Z9-12 Ac" E7Z9-12 Ac  
#> 358  1306 NA            2011   2025 NA    Rapsölmethy… "huile de c… Estere meti…
#> 359  1307 NA            2011   2011 NA    Kunstharz-D… "résine syn… Dispersione…
#> 360  1308 NA            2011   2011 NA    Polyurethan… "résine de … Resina di p…
#> 361  1309 NA            2011   2011 Vase… Vaseline     "vaseline"   Vaselina    
#> 362  1311 NA            2011   2011 4-cy… Dodemorph    "dodémorphe" Dodemorph   
#> 363  1312 NA            2011   2025 Dode… 12 OH (1-do… "12 OH (1-d… 12 OH (1-do…
#> 364  1313 NA            2011   2025 tetr… 14 OH (1-te… "14 OH (1-t… 14 OH (1-te…
#> 365  1314 DD87635F…     2011   2026 Räub… Macrolophus… "Macrolophu… Macrolophus…
#> 366  1315 BC780878…     2011   2026 NA    Microterys … "Microterys… Microterys …
#> 367  1316 360E8FC1…     2011   2026 NA    Pseudaphycu… "Pseudaphyc… Pseudaphycu…
#> 368  1317 634E1B51…     2011   2026 meth… Kresoxim-me… "krésoxim-m… Kresoxim-me…
#> 369  1319 24E6793B…     2011   2026 meth… Azoxystrobin "azoxystrob… Azossistrob…
#> 370  1322 NA            2011   2017 Kali… Kaliseife    "savon de p… Sapone pota…
#> 371  1324 A9BEF2B8…     2011   2026 Raub… Amblyseius … "Amblyseius… Amblyseius …
#> 372  1325 4DB8B083…     2011   2026 Para… Aphidius er… "Aphidius e… Aphidius er…
#> 373  1326 NA            2011   2018 3-(2… Fluquincona… "fluquincon… Fluquincona…
#> 374  1331 7DE422F7…     2011   2026 Equi… Schachtelha… "extrait de… Estratto di…
#> 375  1334 NA            2011   2017 Quar… Quarzsand    "sable de q… Sabbia di q…
#> 376  1335 NA            2011   2017 Styr… Styrol-Buty… "styrol-but… Copolimeri …
#> 377  1336 NA            2011   2011 (RS)… Cypermethri… "cyperméthr… Cipermetrin…
#> 378  1342 NA            2011   2025 Alum… Schwefelsau… "acide sulf… Acido solfo…
#> 379  1343 NA            2011   2013 Benz… Benzyl-Dode… "bromure de… Bromuro di …
#> 380  1344 NA            2011   2022 3-an… Famoxadone   "famoxadone" Famoxadone  
#> 381  1345 NA            2011   2021 5,7-… Quinoxyfen   "quinoxyfen" Quinoxifen  
#> 382  1348 NA            2011   2012 ethy… Triazamat    "triazamate" Triazamate  
#> 383  1349 NA            2011   2024 (S)-… S-Metolachl… "S-métolach… S-Metolaclor
#> 384  1350 D6BBFA61…     2011   2026 (E)-… Acetamiprid  "acétamipri… Acetamiprid 
#> 385  1351 5B80F548…     2011   2026 Para… Eretmocerus… "Eretmoceru… Eretmocerus…
#> 386  1352 25138AA9…     2011   2026 Räub… Chrysoperla… "Chrysoperl… Chrysoperla…
#> 387  1355 NA            2011   2017 Ethy… Cinidon-eth… "cinidon ét… Cinidon-eti…
#> 388  1356 NA            2011   2020 meth… Flupyrsulfu… "flupyrsulf… Flupirsulfu…
#> 389  1359 92A1AF69…     2011   2026 fatt… Fettsäuren   "acides gra… Acidi grassi
#> 390  1360 1EE56990…     2011   2026 Acet… Essigsäure   "acide acét… Acido aceti…
#> 391  1362 NA            2011   2023 (S)-… Indoxacarb   "indoxacarb… Indoxacarb  
#> 392  1363 FDD02E7C…     2011   2026 calc… Prohexadion… "prohexadio… Calcio-proe…
#> 393  1364 216523E4…     2011   2026 (E,E… Trifloxystr… "trifloxyst… Trifloxystr…
#> 394  1365 NA            2011   2011 2-ch… Butafenacil  "butafenaci… Butafenacil 
#> 395  1366 59487F47…     2011   2026 1-Me… Fenhexamid   "fenhexamid… Fenhexamid  
#> 396  1367 6C28B1D5…     2011   2026 8-te… Spiroxamine  "spiroxamin… Spiroxamina 
#> 397  1368 6D80509F…     2011   2026 2’,6… Florasulam   "florasulam… Florasulam  
#> 398  1369 C9003121…     2011   2026 NA    Pseudomonas… "Pseudomona… Pseudomonas…
#> 399  1371 E7563E12…     2011   2026 Räub… Adalia bipu… "Adalia bip… Adalia bipu…
#> 400  1376 0B8785C9…     2011   2026 Coni… Coniothyriu… "Coniothyri… Coniothyriu…
#> 401  1381 NA            2011   2011 Alka… Triton X 114 "Triton X 1… Triton X 114
#> 402  1382 NA            2011   2011 Diok… Triton GR 5  "Triton GR … Triton GR 5 
#> 403  1384 7C72E161…     2011   2026 Ento… Steinernema… "Steinernem… Steinernema…
#> 404  1385 NA            2011   2017 NA    Gelbsenfmehl "farine de … Farina di s…
#> 405  1391 21FA1CA6…     2011   2026 Benz… Benzoesäure  "acide benz… Acido benzo…
#> 406  1392 NA            2011   2020 NA    Mischung au… "mélange d'… Miscela di …
#> 407  1393 NA            2011   2020 NA    Kieselsand   "sable sili… Sabbia sili…
#> 408  1394 NA            2011   2011 NA    Farbstoffe   "colorants"  Coloranti   
#> 409  1395 92A1AF69…     2011   2026 fatt… Fettsäuren   "acides gra… Acidi grassi
#> 410  1396 51208C7E…     2011   2026 oils… Sesamöl raf… "huile de s… Olio di ses…
#> 411  1397 NA            2011   2025 meth… Thiencarbaz… "thiencarba… thiencarbaz…
#> 412  1398 NA            2011   2025 NA    Proteine     "protéines"  Proteine    
#> 413  1399 NA            2011   2021 N-{3… Thiacloprid  "thiaclopri… Tiacloprid  
#> 414  1400 235C3063…     2011   2026 N-(4… Flufenacet   "flufénacet" Flufenacet  
#> 415  1401 NA            2011   2020 Baci… Bacillus su… "Bacillus s… Bacillus su…
#> 416  1402 NA            2011   2018 2-[1… Tepraloxydim "tépraloxyd… Tepraloxidim
#> 417  1403 NA            2011   2017 (+/-… Novaluron    "novaluron"  Novaluron   
#> 418  1404 905CE62C…     2011   2026 3,5-… Zoxamid      "zoxamide"   Zoxamid     
#> 419  1405 FF3881F7…     2011   2026 sodi… Iodosulfuro… "iodosulfur… Iodosulfuro…
#> 420  1406 NA            2011   2025 7-fl… Flumioxazin  "Flumioxazi… Flumiossazi…
#> 421  1407 78C7343C…     2011   2026 2-(4… Etofenprox   "etofenprox" Etofenprox  
#> 422  1408 NA            2011   2025 diet… Mefenpyr-Di… "méfenpyr d… Mefenpir-di…
#> 423  1409 BE5B3348…     2011   2026 (RS)… Fluroxypyr-… "fluroxypyr… Fluroxypyr-…
#> 424  1410 DD442F5B…     2011   2026 ferr… Eisen-III-P… "phospate d… Fosfato di …
#> 425  1411 NA            2011   2011 orth… Chlorophen   "chlorophèn… Clorofene   
#> 426  1412 NA            2011   2011 para… Para-tertiä… "amylphénol… p-ter-amilf…
#> 427  1413 3CAC85F2…     2011   2026 Ento… Heterorhabd… "Heterorhab… Heterorhabd…
#> 428  1414 966E9070…     2011   2026 Ento… Photorhabdu… "Photorhabd… Photorhabdu…
#> 429  1416 BD9791A5…     2011   2026 5-cy… Isoxaflutole "isoxafluto… Isoxaflutole
#> 430  1418 NA            2011   2025 NA    Verbissmitt… "répulsif d… Repellente …
#> 431  1421 D67B0846…     2011   2026 mixt… Spinosad     "spinosad"   Spinosad    
#> 432  1422 B4DB7526…     2011   2026 2-(4… Mesotrione   "mésotrione" Mesotrione  
#> 433  1423 9E7171A1…     2011   2026 S-2-… Dimethenami… "diméthénam… Dimethenami…
#> 434  1425 C69216D9…     2011   2026 meth… Pyraclostro… "pyraclostr… Pyraclostro…
#> 435  1426 NA            2011   2018 N-(2… Metosulam    "métosulame" Metosulam   
#> 436  1428 ED934DE2…     2011   2026 {2-M… Iprovalicarb "iprovalica… Iprovalicarb
#> 437  1430 468E883D…     2018   2026 Calc… Schwefelkal… "polysulfur… zolfo calci…
#> 438  1434 8D5E29C3…     2011   2026 (RS)… Imazamox     "imazamox"   Imazamox    
#> 439  1437 NA            2011   2015 Raub… Hypoaspis a… "Hypoaspis … Hypoaspis a…
#> 440  1439 NA            2011   2025 2-me… Methylbuten… "méthylbute… Metilbuteno…
#> 441  1442 694D1981…     2011   2026 Ento… Xenorhabdus… "Xenorhabdu… Xenorhabdus…
#> 442  1444 2E117D6A…     2011   2026 sodi… Propoxycarb… "propoxycar… Propoxicarb…
#> 443  1447 NA            2011   2015 NA    Netzmittel … "agents ten… Umidificant…
#> 444  1449 NA            2011   2016 NA    Netzmittel … "agents ten… Umidificant…
#> 445  1456 209EF971…     2011   2026 ethy… Pyraflufen-… "pyraflufen… Pyraflufen-…
#> 446  1457 8C577BBF…     2011   2026 1-(4… Foramsulfur… "foramsulfu… Foramsulfur…
#> 447  1458 NA            2011   2023 (RS)… Etoxazol     "etoxazole"  Etoxazol    
#> 448  1459 NA            2011   2021 (S)-… Fenamidon    "fénamidon"  Fenamidone  
#> 449  1460 NA            2011   2020 meth… Picoxystrob… "picoxystro… Picoxistrob…
#> 450  1462 NA            2011   2023 N-te… Methoxyfeno… "méthoxyfén… Metossifeno…
#> 451  1463 NA            2011   2020 (RS)… Beflubutamid "beflubutam… Beflubutamid
#> 452  1464 NA            2011   2025 Ethy… Isoxadifen-… "isoxadifen… Isoxadifen-…
#> 453  1465 73688E44…     2011   2026 Räub… Anthocoris … "Anthocoris… Anthocoris …
#> 454  1466 29E2E104…     2011   2026 meth… Mesosulfuro… "mesosulfur… mesosulfuro…
#> 455  1467 NA            2011   2025 Isop… Benthiavali… "benthiaval… Bentiavalic…
#> 456  1468 F5457E14…     2011   2026 2-Ch… Boscalid     "boscalide"  Boscalid    
#> 457  1469 AC35B7AC…     2011   2026 4-ch… Cyazofamid   "cyazofamid… Ciazofamid  
#> 458  1470 NA            2011   2022 3-(2… Spirodiclof… "spirodiclo… Spirodiclof…
#> 459  1471 NA            2011   2018 1-(2… Clothianidin "clothianid… Clothianidin
#> 460  1472 B8390F01…     2011   2026 Räub… Feltiella a… "Feltiella … Feltiella a…
#> 461  1473 NA            2011   2022 3-(2… Thiamethoxam "thiaméthox… Thiamethoxam
#> 462  1474 NA            2011   2018 (RS)… beta-Cyflut… "beta-cyflu… beta-Ciflut…
#> 463  1475 20260754…     2011   2026 2-ch… Pethoxamid   "pethoxamid… Petoxamide  
#> 464  1476 D9EF28D6…     2011   2026 (RS)… Prothiocona… "prothiocon… Protioconaz…
#> 465  1479 430AD1E9…     2011   2026 Ento… Heterorhabd… "Heterorhab… Heterorhabd…
#> 466  1480 NA            2011   2025 (Z)-… Z9-14Ac      "Z9-14Ac"    Z9-14 Ac    
#> 467  1481 9F229C2C…     2011   2026 alph… Flutolanil   "flutolanil" flutolanil  
#> 468  1482 F4D3BC38…     2011   2026 (E)-… Fluoxastrob… "fluoxastro… Fluoxastrob…
#> 469  1483 C2A84945…     2011   2026 NA    Polyalkylen… "polyalkyle… Polyalkylen…
#> 470  1486 12B1143D…     2011   2026 Para… Eretmocerus… "Eretmoceru… Eretmocerus…
#> 471  1487 NA            2011   2025 octa… E3Z13-18Ac   "E3Z13-18Ac" E3Z13-18Ac  
#> 472  1488 NA            2011   2025 NA    Magermilchp… "lait écrém… latte magro…
#> 473  1489 NA            2011   2011 2,6,… alpha-Pinen  "alpha-pinè… alfa-Pinene 
#> 474  1490 NA            2011   2011 4-Is… Limonen      "limonène"   Limonen     
#> 475  1491 NA            2011   2011 NA    Wollwachs (… "graisse de… Cera di lan…
#> 476  1492 NA            2011   2025 octa… E2Z13-18Ac   "E2Z13-18Ac" E2Z13-18Ac  
#> 477  1494 4BE0D277…     2011   2026 3-do… Acequinocyl  "acéquinocy… Acechinocil 
#> 478  1496 NA            2011   2020 NA    Eukalyptusöl "huile d'eu… Olio di euc…
#> 479  1497 NA            2011   2020 NA    Pfefferminz… "Huile de m… olio di men…
#> 480  1498 NA            2011   2016 NA    Codlemone    "codlemone"  Codlemone   
#> 481  1500 BEF27B58…     2011   2026 6-io… Proquinazid  "proquinazi… Proquinazid 
#> 482  1501 666A6D3F…     2011   2026 1-me… 1-methylcyc… "1-méthylcy… 1-metilcicl…
#> 483  1503 NA            2011   2020 NA    Wintergrünöl "huile de g… Olio di gau…
#> 484  1506 NA            2011   2011 NA    S-CIS-Verbe… "S-cis-verb… (S)-cis-ver…
#> 485  1507 28F3C20A…     2011   2026 (Z)-… Cyflufenamid "cyflufenam… Ciflufenamid
#> 486  1508 2A0BCAA6…     2011   2026 1-(4… Tritosulfur… "tritosulfu… Tritosulfur…
#> 487  1509 81609369…     2011   2026 3'-B… Metrafenone  "métrafénon… Metrafenone 
#> 488  1510 0B26FA4D…     2011   2026 Nona… Pelargonsäu… "acide péla… Acido pelar…
#> 489  1517 NA            2011   2025 (Z,Z… Z3Z13-18Ac   "Z3Z13-18AC" Z3Z13-18AC  
#> 490  1522 8A57674A…     2011   2026 Para… Trichogramm… "Trichogram… Trichogramm…
#> 491  1526 1FB80A69…     2024   2026 Raub… Amblyseius … "Amblyseius… Amblyseius …
#> 492  1527 D93AAC59…     2011   2026 (6R,… Milbemectin  "milbemecti… Milbemectin 
#> 493  1528 CCA4A392…     2011   2026 2,6-… Fluopicolide "fluopicoli… Fluopicolide
#> 494  1529 EAE2CCDA…     2011   2026 8-(2… Pinoxaden    "pinoxaden"  pinoxaden   
#> 495  1531 8EAF59C3…     2011   2026 Alum… Kaolin       "caolin"     Caolino     
#> 496  1532 NA            2011   2025 pota… Kaliumhydro… "Hydrogénca… Idrogencarb…
#> 497  1533 4B0F1FD6…     2011   2026 (1?3… Laminarin    "laminarin"  Laminarin   
#> 498  1535 90F47BB6…     2011   2026 NA    Amblyseius … "Amblyseius… Amblyseius …
#> 499  1537 NA            2011   2017 pota… Kaliumiodid  "iodure de … Ioduro di p…
#> 500  1538 NA            2011   2017 pota… Kaliumthioc… "thiocyanat… Tiocianato …
#> 501  1539 03E43355…     2011   2026 NA    Schafsfett   "Gras de mo… Grasso di p…
#> 502  1540 NA            2011   2019 NA    Azadirachti… "azadiracht… Azadiractin…
#> 503  1541 CDA8D78B…     2011   2026 meth… Benalaxyl-M  "bénalaxyl-… Benalaxil-M 
#> 504  1542 9D73C4BB…     2011   2026 N6-b… 6-benzylade… "6-benzylad… 6-benzilade…
#> 505  1543 7500A9E0…     2011   2026 N-cy… Flonicamid   "Flonicamid… Flonicamide 
#> 506  1545 NA            2011   2017 5-te… Oxadiargyl   "Oxadiargyl" Oxadiargyl  
#> 507  1546 5BF729C3…     2011   2026 2-{2… Tembotrione  "tembotrion… tembotrione 
#> 508  1547 F27EE7FE…     2011   2026 Aure… Aureobasidi… "Aureobasid… Aureobasidi…
#> 509  1548 33016EA8…     2011   2026 Para… Trichogramm… "Trichogram… Trichogramm…
#> 510  1549 ED53FB18…     2011   2026 NA    Habrobracon… "Habrobraco… Habrobracon…
#> 511  1550 BD699D51…     2011   2026 Para… Lariophagus… "Lariophagu… Lariophagus…
#> 512  1552 33778929…     2011   2026 1,1-… Mepiquatchl… "Mepiquat-c… mepiquat cl…
#> 513  1554 3016B169…     2011   2026 2-(4… Mandipropam… "Mandipropa… Mandipropam…
#> 514  1556 9E8499AE…     2011   2026 N-[4… Cyprosulfam… "cyprosulfa… cyprosulfam…
#> 515  1560 E166875A…     2014   2026 3-(3… Amisulbrom   "Amisulbrom" Amisulbrom  
#> 516  1562 7A17715D…     2011   2026 sulf… Sulfurylflu… "Sulfuryl F… Floruro di …
#> 517  1563 336AC084…     2025   2026 Triz… Zinkphosphid "Phosphure … Fosfuro di …
#> 518  1564 BAA1BA06…     2011   2026 N-(5… Pyroxsulam   "Pyroxsulam" Pyroxsulam  
#> 519  1565 CCD597BE…     2011   2026 Raub… Typhlodromi… "Typhlodrom… Typhlodromi…
#> 520  1566 NA            2011   2025 NA    Polyethermo… "polyéther … trisiloxane…
#> 521  1567 2312C520…     2011   2026 isop… Bifenazat    "bifénazate" bifenazato  
#> 522  1568 EE8F4C48…     2011   2026 cis-… Spirotetram… "spirotetra… spirotetram…
#> 523  1569 78948E5C…     2013   2026 NA    Anisopterom… "Anisoptero… Anisopterom…
#> 524  1570 EA526877…     2015   2026 4-am… Aminopyralid "aminopyral… aminopyralid
#> 525  1571 NA            2011   2025 Anio… Synthetisch… "Latex synt… Lattice sin…
#> 526  1572 NA            2011   2025 NA    Xanthan      "xanthane"   xanthan     
#> 527  1573 NA            2011   2025 NA    Alkylpolygl… "alkyl poly… alchil poli…
#> 528  1574 5C0BFB4B…     2011   2026 (4’’… Emamectinbe… "émamectine… emamectina …
#> 529  1575 373C1C2E…     2011   2026 3-br… Chlorantran… "chlorantra… chlorantran…
#> 530  1576 NA            2011   2020 NA    Naphtensäur… "Acides nap… Acidi nafte…
#> 531  1578 NA            2011   2020 NA    Montanwachs… "Montanate"  Montanate   
#> 532  1579 D5D9B1C2…     2015   2026 meth… Valifenalate "Valifenala… Valifenalate
#> 533  1581 958E6100…     2011   2026 N-(3… Bixafen      "Bixafen"    Bixafen     
#> 534  1583 68FB63B0…     2013   2026 N-{2… Fluopyram    "Fluopyram"  Fluopyram   
#> 535  1584 BD01EED4…     2011   2026 NA    Pseudomonas… "Pseudomona… Pseudomonas…
#> 536  1585 DB8FB148…     2014   2026 Heli… Baumwollkap… "Helicoverp… Helicoverpa…
#> 537  1588 384C3717…     2018   2026 (2R,… Spinetoram   "spinetoram" spinetoram  
#> 538  1589 5F235D17…     2017   2026 (RS)… Penthiopyrad "Penthiopyr… Penthiopyrad
#> 539  1593 E74C6EF2…     2015   2026 3-(2… Penoxsulam   "penoxsulam" penoxsulam  
#> 540  1598 17C5DDF6…     2014   2026 NA    Purpureocil… "Purpureoci… Purpureocil…
#> 541  1656 9BCD43E0…     2014   2026 NA    Gliocladium… "Gliocladiu… Gliocladium…
#> 542  1659 CA9AE949…     2019   2026 Oran… Orangenöl    "Huile d'or… Olio d'aran…
#> 543  1660 740DC585…     2014   2026 Malt… Maltodextrin "maltodextr… maltodestri…
#> 544  1661 B325A5DF…     2014   2026 S-al… Fenpyrazamin "fenpyrazam… fenpyrazami…
#> 545  1662 B57E6C3E…     2015   2026 Deca… 1-Decanol    "1-décanol"  1-Decanol   
#> 546  1672 NA            2011   2026 NA    Docusatnatr… "docusate s… docusato so…
#> 547  1689 NA            2011   2026 NA    1,2-Benziso… "1,2-benzis… 1,2-benziso…
#> 548  1701 3E5EEF6E…     2011   2026 Alum… Aluminiumsu… "Sulfate d'… Solfato di …
#> 549  1733 NA            2011   2026 NA    Polyglykole… "Ether de p… Etere di po…
#> 550  1739 541A8565…     2015   2026 3-(d… Fluxapyroxad "fluxapyrox… Fluxapyroxa…
#> 551  1740 28F33B73…     2019   2026 Alum… Kaliumalumi… "Sulfate d'… Solfato di …
#> 552  1744 AE92F698…     2015   2026 Para… Ephedrus ce… "Ephedrus c… Ephedrus ce…
#> 553  1746 B332EE66…     2015   2026 Para… Praon voluc… "Praon volu… Praon voluc…
#> 554  1748 38DDF6FB…     2017   2026 NA    Bacillus am… "Bacillus a… Bacillus am…
#> 555  1751 NA            2013   2022 (±)-… Propiconazo… "propiconaz… Propiconazo…
#> 556  1753 NA            2016   2025 2-ch… Chlormequat  "Chlorméqua… Clormequat  
#> 557  1755 6C0978A8…     2018   2026 Ethen Ethylen      "Éthylène"   Etilene     
#> 558  1756 NA            2018   2025 calc… Calciumcarb… "Carbonate … Carbonato d…
#> 559  1760 0397246E…     2018   2026 Line… COS-OGA      "COS-OGA"    COS-OGA     
#> 560  1762 1CB53F47…     2015   2026 (Z)-… Z11-14OH     "Z11-14OH"   Z11-14OH    
#> 561  1776 NA            2015   2025 NA    Sodium Laur… "Sodium Lau… Sodium Laur…
#> 562  1789 2DB1F9EF…     2015   2026 L-Ca… Grüne-Minze… "huile de m… olio di men…
#> 563  1793 7867F638…     2022   2026 meth… Halauxifen-… "Halauxifen… Halauxifen-…
#> 564  1795 F47B3C0E…     2019   2026 N-[9… Benzovindif… "benzovindi… benzovindif…
#> 565  1800 8381D435…     2019   2026 5-et… Ametoctradin "Ametoctrad… Ametoctradi…
#> 566  1806 NA            2016   2025 NA    Anionische … "Tensioacti… Tensioattiv…
#> 567  1807 NA            2016   2025 NA    Nichtionisc… "Tensioacti… Tensioattiv…
#> 568  1815 6C023B02…     2019   2026 NA    Pepinomosai… "virus de l… virus del m…
#> 569  1844 NA            2016   2025 NA    Polyethylen… "Polyethyle… Polyethylen…
#> 570  1845 A5518717…     2022   2026 diso… Dinatriumph… "phosphonat… fosfonato d…
#> 571  1846 NA            2016   2026 NA    2-hydroxy-4… "2-hydroxy-… 2-hydroxy-4…
#> 572  1863 NA            2016   2025 NA    Fettalkohol… "Ether poly… alcol grass…
#> 573  1864 NA            2016   2025 NA    Ethoxyliert… "éthoxylate… olio di ric…
#> 574  1875 178335E8…     2016   2026 Fatt… Fettsäuren … "Acides gra… Acidi grass…
#> 575  1876 NA            2017   2026 NA    Naphthalene… "Naphthalen… Naphthalene…
#> 576  1877 23CE0251…     2021   2026 1,4-… 1,4-Dimethy… "1,4-diméth… 1,4-dimetil…
#> 577  1878 C53F59C1…     2023   2026 (5-c… Pyriofenon   "Pyriofenon… Pyriofenone 
#> 578  1887 NA            2025   2025 NA    Saccharomyc… "Saccharomy… Saccharomyc…
#> 579  1911 NA            2017   2025 NA    Terpineol    "Terpineol"  Terpineol   
#> 580  1922 NA            2024   2025 NA    Kupfer (als… "cuivre (so… Rame (sotto…
#> 581  1927 NA            2021   2025 NA    Fettsäuren … "Acides gra… Acidi grass…
#> 582  1928 NA            2022   2025 NA    Ethoxyliert… "Huile de s… Olio di soi…
#> 583  1933 NA            2025   2025 (2RS… Mefentriflu… "Mefentrifl… Mefentriflu…
#> 584  1936 NA            2025   2025 fish… Fischöl      "Huile de p… Olio di pes…
#> 585  1939 FE5D35CF…     2020   2026 Baci… Bacillus am… "Bacillus a… Bacillus am…
#> 586  1954 9082E5DC…     2025   2026 Para… Encarsia ci… "Encarsia c… Encarsia ci…
#> 587  1956 491AFFA8…     2023   2026 Räub… Rodolia car… "Rodolia ca… Rodolia car…
#> 588  1957 CB64F094…     2019   2026 GA4:… Gibberellin  "gibberilli… Gibberellina
#> 589  1964 NA            2023   2025 NA    Hydroxyprop… "amidon hyd… amido idros…
#> 590  1974 64FB7C16…     2023   2026 Para… Coccophagus… "Coccophagu… Coccophagus…
#> 591  1977 NA            2025   2025 Para… Anagyrus vl… "Anagyrus v… Anagyrus vl…
#> 592  1978 D3239149…     2024   2026 Para… Trichopria … "Trichopria… Trichopria …
#> 593  1979 CC9603D7…     2024   2026 Para… Cephalonomi… "Cephalonom… Cephalonomi…
#> 594  1990 9ACEF16A…     2024   2026 Räub… Eupeodes co… "Eupeodes c… Eupeodes co…
#> 595  1993 ED00F27E…     2024   2026 Ento… Heterorhabd… "Heterorhab… Heterorhabd…
#> 596  1999 B6CB6484…     2024   2026 Para… Anastatus b… "Anastatus … Anastatus b…
#> 597  2000 85D6C935…     2023   2026 Raub… Transeius m… "Transeius … Transeius m…
#> 598  2001 4CD94ADD…     2024   2026 Räub… Sphaerophor… "Sphaeropho… Sphaerophor…
#> 599  2012 544BE57D…     2022   2026 (3S,… Gibberellin… "Acide gibb… Acido gibbe…
#> 600  2016 99E6979D…     2025   2026 Räub… Exochomus q… "Exochomus … Exochomus q…
#> 601  2019 B3F771AC…     2025   2026 Para… Trissolcus … "Trissolcus… Trissolcus …
#> 602  2023 NA            2023   2025 6,6-… Terpen-Olig… "Oligomères… Oligomeri t…
#> 603  2024 9E05236A…     2025   2026 Raub… Pronematus … "Pronematus… Pronematus …
#> 604  2042 NA            2025   2025 NA    Kieselgur (… "Kieselgur … Kieselgur (…
#> 605    NA 028AA985…     2026   2026 (9Z)… (Z)-9-Dodec… "(Z)-9-Dode… NA          
#> 606    NA 057FC3E0…     2026   2026 sodi… MCPB Na      "MCPB sodiu… NA          
#> 607    NA 0974AB08…     2026   2026 [(3,… Triclopyrtr… "triclopyr-… NA          
#> 608    NA 0A7BFE30…     2026   2026 (E)-… (E)-8-Dodec… "(E)-8-dode… NA          
#> 609    NA 0DEFCF91…     2026   2026 Fish… Repellents … "Répulsifs … NA          
#> 610    NA 13D3F11A…     2026   2026 1-(4… Oxathiapipr… "Oxathiapip… NA          
#> 611    NA 1D7FC783…     2026   2026 Whit… Paraffinöl   "huile de p… NA          
#> 612    NA 215D748A…     2026   2026 Ento… Apfelwickle… "virus de l… NA          
#> 613    NA 26CA7D7E…     2026   2026 meth… thiencarbaz… "thiencarba… NA          
#> 614    NA 2862E3B1…     2026   2026 copp… Dreibasisch… "Cuivre (so… NA          
#> 615    NA 306B272F…     2026   2026 Para… Anagyrus vl… "Anagyrus v… NA          
#> 616    NA 32E98251…     2026   2026 Dode… Dodecan-1-ol "Dodecan-1-… NA          
#> 617    NA 332FDA27…     2026   2026 (7E,… (E,Z)-7,9-D… "(E,Z)-7,9-… NA          
#> 618    NA 3788E18C…     2026   2026 Diat… Kieselgur (… "Kieselgur … NA          
#> 619    NA 37FEF947…     2026   2026 (Z)-… (Z)-8-dodec… "(Z)-8-Dodé… NA          
#> 620    NA 47115E42…     2026   2026 (Z)-… (Z)-9-Tetra… "(Z)-9-Tetr… NA          
#> 621    NA 4753372E…     2026   2026 (E,E… E8E10-12 OH… "(E,E)-8,10… NA          
#> 622    NA 50804EF0…     2026   2026 lime… Kalkstein    "calcaire"   NA          
#> 623    NA 51035D33…     2026   2026 Calc… Calciumcarb… "Carbonate … NA          
#> 624    NA 5226F9F7…     2026   2026 3,5,… Triclopyrbu… ""           NA          
#> 625    NA 6D40EA6E…     2026   2026 7-fl… Flumioxazin  "flumioxazi… NA          
#> 626    NA 738E64D8…     2026   2026 (Z)-… (Z)-8-Dodec… "(Z)-8-Dode… NA          
#> 627    NA 7468DFE9…     2026   2026 Not … Blutmehl     "farine de … NA          
#> 628    NA 7F0BC66A…     2026   2026 (Z)-… Z8-12Ac      "(Z)-8-Dode… NA          
#> 629    NA 86470187…     2026   2026 2-ch… Chlormequat… "chlormequa… NA          
#> 630    NA 92CDFC85…     2026   2026 (2RS… Mefentriflu… "Mefentrifl… NA          
#> 631    NA 93DF970D…     2026   2026 3,5-… Propyzamide  "propyzamid… NA          
#> 632    NA 96054844…     2026   2026 (E,Z… (E,Z)-2,13-… "(E,Z)-2,13… NA          
#> 633    NA 9B6470F1…     2026   2026 dico… Kupferoxych… "cuivre (so… NA          
#> 634    NA A9525EF1…     2026   2026 copp… Kupferhydro… "cuivre (so… NA          
#> 635    NA B349A6C9…     2026   2026 (2,4… 2,4-D-2-Eth… ""           NA          
#> 636    NA BCA8A28C…     2026   2026 copp… Kupferoctan… "cuivre (so… NA          
#> 637    NA BE6984AA…     2026   2026 n-Do… n-Dodecyl a… "n-Dodecyl … NA          
#> 638    NA BEB1622C…     2026   2026 (E,Z… (E,Z,Z)-3,8… "(E,Z,Z)-3,… NA          
#> 639    NA C1966A3B…     2026   2026 tetr… n-Tetredecy… "n-Tetradec… NA          
#> 640    NA C9775705…     2026   2026 Fatt… Fettsäuren,… ""           NA          
#> 641    NA CC42B743…     2026   2026 (E,Z… (E,Z)-3,13-… "(E,Z)-3,13… NA          
#> 642    NA CFC415F9…     2026   2026 (Z,Z… (Z,Z)-3,13-… "(Z,Z)-3,13… NA          
#> 643    NA DC2C6844…     2026   2026 (E,Z… (E,Z)-3,8-T… "(E,Z)-3,8-… NA          
#> 644    NA DE9DB824…     2026   2026 NA    Bacillus am… "Bacillus a… NA          
#> 645    NA E6834FD5…     2026   2026 Tetr… Tetradecan-… "Tetradecan… NA          
#> 646    NA EE073DEB…     2026   2026 Pota… Kaliumhydro… "Hydrogogén… NA          
#> 647    NA F2EC442B…     2026   2026 (Z)-… (Z)-11-Tetr… "(Z)-11-Tet… NA          
#> 648    NA F507F176…     2026   2026 NA    Saccharomyc… "Saccharomy… NA          
#> # ℹ 1 more variable: substance_en <chr>